requirements.txt
Flask==2.0.1
pytesseract==0.3.8
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
